DagasServer/requirements.txt
Django~=3.2.10
djangorestframework~=3.12.4
django-rest-auth~=0.9.5
qrcode~=7.3.1